don't recreate Gun's Image every frame
Instead of storing resource URLs for horizontal and vertical in Gun and
allocating and assigning a new Image to this.img every frame, create two
Images, store both in Gun, and choose which to draw each frame.

This eliminates tons of allocation messages (one per frame) on the console, and
should be more efficient.